fix margins on how it works page
